(SC-6) Renamed file and classes name to match them, deleted comment, deleted a grid
