+++ src/components/champion/index.tsx
@@ -1,15 +1,10 @@
 import * as React from "react";
-import { Grid } from "@material-ui/core";
-import { Selector } from "./selector";
+import { ChampionSelector } from "./ChampionSelector";
 
 export class ChampionPage extends React.Component {
     render() {
         return (
-            <Grid container spacing={3}>
-                <Grid item xs>
-                    <Selector  />
-                </Grid>
-            </Grid>
+            <ChampionSelector  />
         )
     }
 }
